extract default command
