+++ validator/src/commands/default/mod.rs
@@ -1,0 +1,4 @@
+pub mod args;
+pub mod execute;
+
+pub use {args::add_args, execute::execute};
