+++ validator/src/commands/mod.rs
@@ -1,5 +1,6 @@
 pub mod authorized_voter;
 pub mod contact_info;
+pub mod default;
 pub mod exit;
 pub mod monitor;
 pub mod plugin;
